Moved stuff more out into csrankings.py.
